requirements.txt
absl-py==2.1.0
altgraph==0.17.4
asgiref==3.8.1
astunparse==1.6.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
distlib==0.3.9
docutils==0.21.2
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
future==1.0.0
gast==0.6.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==1.0.0
mysql-connector-python==9.1.0
namex==0.0.8
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
plyer==2.1.0
protobuf==5.29.0
ratelim==0.1.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tk==0.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.1
Werkzeug==3.1.3
wrapt==1.17.0